backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.6.1
httpx>=0.27.0
httpcore>=0.18.0 